feat: componentization hide button
